Roll Verifier Updates
Roll Verifier updates to bunch of sites

+++ DiceBot/DiceSite.cs
@@ -215,7 +215,7 @@ namespace DiceBot
             }
             return 0;
         }
-        public static decimal sGetLucky(string server, string client, int nonce)
+        public static decimal sGetLucky(string server, string client, long nonce)
         {
             HMACSHA512 betgenerator = new HMACSHA512();
