+++ DiceBot/WinDice.cs
@@ -205,7 +205,7 @@ namespace DiceBot
             
         }
 
-        public static new decimal sGetLucky(string server, string client, int nonce)
+        public static new decimal sGetLucky(string server, string client, long nonce)
         {
             SHA512 betgenerator = SHA512.Create();
