+++ DiceBot/betterbets.cs
@@ -443,7 +443,7 @@ namespace DiceBot
             }
             return 0;
         }
-        new public static decimal sGetLucky(string server, string client, int nonce)
+        new public static decimal sGetLucky(string server, string client, long nonce)
         {
             HMACSHA512 betgenerator = new HMACSHA512();
