+++ DiceBot/bitdice.cs
@@ -233,7 +233,7 @@ namespace DiceBot
             }
 
         }
-        public static new decimal sGetLucky(string server, string client, int nonce)
+        public static new decimal sGetLucky(string server, string client, long nonce)
         {
             HMACSHA512 betgenerator = new HMACSHA512();
